Add Makefile target "man-html" for rebuilding the (checked-in) HTML versions of the man pages.  This allows you to view the changes with "svn diff" before running "make upload-man" which commits them.
Reviewed by Craig Silverstein.

+++ changes/fergus.henderson/r6s3/trunk/src/include_server_if.c
@@ -19,6 +19,8 @@
 
 /* Authors: Manos Renieris, Fergus Henderson */
 
+#include <config.h>
+
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
@@ -83,8 +85,8 @@ int dcc_talk_to_include_server(char **argv, char ***files)
     }
 
     if (strlen(include_server_port) >= ((int)sizeof(sa.sun_path) - 1)) {
-        rs_log_warning("$INCLUDE_SERVER_PORT is longer than %d characters",
-                       (sizeof(sa.sun_path) - 1));
+        rs_log_warning("$INCLUDE_SERVER_PORT is longer than %ld characters",
+                       ((long) sizeof(sa.sun_path) - 1));
         return 1;
     }
